misc/_Source_eBook_DevOps/devops_code_book/chapter16/otp_dock/python/requirements.txt
redis
nameko
validate_email
python-dotenv 
